Day 07: Added Game Stages and Working Menu Buttons

+++ src/entites/Entity.java
@@ -2,7 +2,6 @@ package entites;
 
 import java.awt.Color;
 import java.awt.Graphics;
-import java.awt.Rectangle;
 import java.awt.geom.Rectangle2D;
 
 public abstract class Entity {
@@ -24,7 +23,7 @@ public abstract class Entity {
 
 	}
 
-	protected void initHitbox(float x, float y, float width, float height) {
+	protected void initHitbox(float x, float y, int width, int height) {
 		hitbox = new Rectangle2D.Float(x, y, width, height);
 	}
